requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-docx==1.1.2
reportlab==4.0.7
pdfplumber==0.11.4
anthropic==0.50.0
playwright==1.51.0
aiofiles==24.1.0
openai==1.58.1
email-validator==2.2.0
bcrypt==4.2.1
cryptography==42.0.8
slowapi==0.1.9
filetype==1.2.0
pyotp==2.9.0
qrcode==8.0
httpx==0.27.0
firecrawl-py==4.12.0
aiohttp==3.11.11
beautifulsoup4==4.12.3
resend==2.5.0
